requirements.txt
tensorflow
opencv-python
numpy
pillow
matplotlib
lxml
streamlit
protobuf
